Move logic to make work with setuptools
