+++ pyproject.toml
@@ -49,8 +49,8 @@ build-backend = "setuptools.build_meta"
 #    "myst-parser==4.0.0",
 #]
 
-[project.entry-points."apache_airflow_provider"]
-provider_info = "datarobot_provider.__init__:get_provider_info"
+#[project.entry-points."apache_airflow_provider"]
+#provider_info = "datarobot_provider.__init__:get_provider_info"
 
 [tool.ruff]
 # General settings
