+++ setup.py
@@ -12,10 +12,10 @@
 from setuptools import find_packages
 from setuptools import setup
 
-from setup_common import DESCRIPTION_TEMPLATE
-from setup_common import classifiers
-from setup_common import common_setup_kwargs
-from setup_common import version
+from datarobot_provider._setup import DESCRIPTION_TEMPLATE
+from datarobot_provider._setup import classifiers
+from datarobot_provider._setup import common_setup_kwargs
+from datarobot_provider._setup import version
 
 python_versions = ">= 3.9"
